refactor(backend): colocate source-owned tests
