+++ README.md
@@ -15,6 +15,7 @@ The backend now lives in `backend/` and keeps the boundaries explicit:
 - `backend/src/presentation` for HTTP, CLI, and MCP entrypoints
 
 Most tests run against in-memory adapters. A small contract suite also runs against the SQL-backed repositories.
+Backend tests use a hybrid layout: source-owned unit tests live beside the code in `backend/src/**/*.test.ts`, while shared support, contract, and integration suites stay centralized under `backend/test`.
 
 ## Commands
