+++ backend/knip.jsonc
@@ -8,6 +8,7 @@
     "src/presentation/mcp/stdio-main.ts!",
     "src/presentation/mcp/http-main.ts!",
     "src/presentation/mcp/miniflare.worker.ts",
+    "src/**/*.test.ts",
     "test/**/*.test.ts",
   ],
