+++ backend/vitest.config.ts
@@ -3,7 +3,7 @@ import { defineConfig } from "vitest/config";
 export default defineConfig({
   test: {
     environment: "node",
-    include: ["test/**/*.test.ts"],
+    include: ["src/**/*.test.ts", "test/**/*.test.ts"],
     setupFiles: ["./vitest.setup.ts"],
   },
 });
